Seperated out feeder functions into a seperate subsystem in order to use the feeder in the autonomous command group (needs to be retested)
