+++ src/main/java/frc/robot/commands/AutonomousCommands/AutonomousRoutine.java
@@ -7,7 +7,7 @@ package frc.robot.commands.AutonomousCommands;
 import edu.wpi.first.wpilibj2.command.ParallelDeadlineGroup;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.RobotContainer;
-import frc.robot.subsystems.AutoFeeder;
+import frc.robot.subsystems.Feeder;
 import frc.robot.subsystems.DriveTrain;
 import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Shooter;
@@ -18,11 +18,11 @@ import frc.robot.subsystems.Shooter;
 public class AutonomousRoutine extends SequentialCommandGroup {
   /** Creates a new AutonomousRoutine. */
   DriveTrain ar_driveTrain;
-  AutoFeeder ar_feeder;
+  Feeder ar_feeder;
   Intake ar_intake;
   Shooter ar_shooter;
 
-  public AutonomousRoutine(DriveTrain drive, AutoFeeder feed, Shooter shoot, Intake intake) {
+  public AutonomousRoutine(DriveTrain drive, Feeder feed, Shooter shoot, Intake intake) {
     ar_feeder = feed;
     ar_shooter = shoot;
     ar_driveTrain = drive; 
@@ -31,7 +31,7 @@ public class AutonomousRoutine extends SequentialCommandGroup {
     // addCommands(new FooCommand(), new BarCommand());
     addCommands( 
         
-      new ParallelDeadlineGroup(new AutoShootRoutine(0.50, 4, ar_shooter), new AutoIntakeRoutine(0.25, 2, ar_intake, ar_feeder)),
+      new ParallelDeadlineGroup(new AutoShootRoutine(0.50, 4, ar_shooter), new AutoIntakeRoutine(0.5, 10, ar_feeder)),
       new DriveDistance(24, 0.50, ar_driveTrain)
